+++ src/converter/computed-style-snapshot.browser.js
@@ -29,6 +29,7 @@
     "fontSize",
     "fontWeight",
     "fontStyle",
+    "textTransform",
     "marginTop",
     "marginRight",
     "marginBottom",